perf: Avoid sorting and deduplicating in Node.flatten (#31)
With `Vec` in place, flattening takes a merge sort, which is `O(n*log(n))`.
Then, an additional pass to deduplicate O(n) is required. This can be avoided
by using a HashSet, which is guaranteed to be O(1) on insert. Also,
a HashSet doesn't need any intermediate memory allocations.